backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
google-generativeai==0.7.2
pydantic==2.6.3
python-dotenv==1.0.1
pypdf==4.0.1
python-docx==1.1.0
supabase==2.3.7
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
openai>=1.10.0
httpx>=0.25.0
